requirements.txt
flask==3.0.0
Werkzeug==3.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-socketio==5.3.6
flask-migrate==4.0.5
flask-cors==4.0.0
flask-talisman==1.1.0
flask-mail==0.9.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-Dance==7.0.0
gevent==23.9.1
gevent-websocket==0.10.1
gunicorn==21.2.0
eventlet==0.33.3
waitress==2.1.2
websockets==14.2
greenlet==3.0.1
google-genai==1.2.0
google-auth==2.29.0
groq==0.9.0
pydantic>=2.5.2
pydantic-core>=2.14.5
httpx>=0.25.2
httpcore>=1.0.2
requests==2.31.0
beautifulsoup4==4.12.2
urllib3>=2.1.0
authlib==1.3.0
itsdangerous==2.1.2
cryptography>=41.0.7
pyopenssl>=23.3.0
moviepy==1.0.3
imageio==2.33.0
imageio-ffmpeg==0.4.9
Pillow>=10.1.0
numpy>=1.23.5
edge-tts==6.1.9
gtts==2.5.0
SpeechRecognition==3.10.0
pydub==0.25.1
pypdf==3.17.1
python-docx==1.1.0
python-pptx==0.6.23
lxml==4.9.3
pytz
blinker
click
typing-extensions>=4.11.0
idna
certifi
six